+++ Mods ChoGGi/Logos/DHARMAInitiativeLogos/metadata.lua
@@ -1,0 +1,23 @@
+return PlaceObj("ModDef", {
+	"title", "DHARMA Initiative Logos",
+	"id", "ChoGGi_DHARMAInitiativeLogos",
+	"steam_id", "3271298123",
+	"pops_any_uuid", "e4384443-837f-48fc-9a4a-fd8197aceac8",
+	"lua_revision", 1007000, -- Picard
+	"version", 1,
+	"version_major", 0,
+	"version_minor", 1,
+	"image", "Preview.png",
+	"author", "ChoGGi & fnix",
+	"code", {
+		"Code/Script.lua",
+	},
+	"has_options", true,
+	"TagCosmetics", true,
+"description", [[
+22 DHARMA Initiative logos from the TV series Lost.
+
+
+Requested by fnix (+ logos supplied)
+]],
+})