+++ Mods ChoGGi/Psychiatric Hospital/metadata.lua
@@ -4,9 +4,9 @@ return PlaceObj("ModDef", {
 	"steam_id", "2923594393",
 	"pops_any_uuid", "b6646233-f947-4d02-8431-11aabcdbe354",
 	"lua_revision", 1007000, -- Picard
-	"version", 2,
+	"version", 3,
 	"version_major", 0,
-	"version_minor", 2,
+	"version_minor", 3,
 	"image", "Preview.jpg",
 	"author", "ChoGGi &  hlechner",
 	"code", {
@@ -17,7 +17,7 @@ return PlaceObj("ModDef", {
 	"description", [[
 Specialized building to treat Idiots.
 
-Mod option to cure Whiners and Introverts.
+Mod options to cure Whiners, Introverts, and Religious.
 
 
 https://github.com/ChoGGi/SurvivingMars_CheatMods/issues/56