+++ Mods ChoGGi/Soylent Glutton Provides Double/metadata.lua
@@ -1,5 +1,5 @@
 return PlaceObj("ModDef", {
-	"title", "Solylent Glutton Provides Double",
+	"title", "Soylent Glutton Provides Double",
 	"id", "ChoGGi_SolylentGluttonProvidesDouble",
 	"steam_id", "1907034997",
 	"pops_any_uuid", "32c54f4c-6008-4644-882c-5cc8403a08a9",
@@ -13,7 +13,7 @@ return PlaceObj("ModDef", {
 		"Code/Script.lua",
 	},
 	"description", [[
-If you have the solylent green breakthrough; gluttons will provide double the soylent.
+If you have the Soylent Green breakthrough; gluttons will provide double the soylent.
 
 
 Not exactly requested by Monovigma...